# oapispec/swagger.yaml
basePath: /
definitions:
  admin.AdminCreateDomainRequest:
    properties:
      autoJoin:
        type: boolean
      autoJoinRoleId:
        type: string
      autoJoinRoleName:
        type: string
      domain:
        type: string
    required:
    - domain
    type: object
  admin.AdminCreateDomainResponse:
    properties:
      domain:
        $ref: '#/definitions/db.OrgDomain'
      message:
        type: string
      success:
        type: boolean
    type: object
...
  admin.AdminDeleteDomainResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
...
  admin.AdminGetDomainVerifyCodeResponse:
    properties:
      code:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
  admin.AdminUpdateDomainRequest:
    properties:
      autoJoin:
        type: boolean
      autoJoinRoleId:
        type: string
      autoJoinRoleName:
        type: string
    type: object
  admin.AdminUpdateDomainResponse:
    properties:
      domain:
        $ref: '#/definitions/db.OrgDomain'
      message:
        type: string
      success:
        type: boolean
    type: object
...
  admin.AdminVerifyDomainResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
      verified:
        type: boolean
    type: object
...
  db.OrgDomain:
    properties:
      autoJoin:
        type: boolean
      autoJoinRoleId:
        type: string
      autoJoinRoleName:
        type: string
      createdAt:
        type: string
      domain:
        type: string
      orgId:
        type: string
      updatedAt:
        type: string
      verified:
        type: boolean
      verifiedAt:
        type: string
    type: object
...
  handlers.VerifyEmailTokenRequest:
    properties:
      token:
        type: string
    required:
    - token
    type: object
  handlers.VerifyEmailTokenResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
...
  reserr.ErrorResponse:
    properties:
      debug:
        description: DebugMessage is a technical message that can be used for debugging.
        type: string
      message:
        description: Message is a user-friendly message that can be displayed to the
          end user.
        type: string
    type: object
...
host: localhost:3360
info:
  contact:
    email: contact@nbrglm.com
    name: NBRGLM Support
    url: https://nbrglm.com/support
  description: This is the NBRGLM Nexeres API Spec documentation.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://nbrglm.com/nexeres/terms
  title: NBRGLM Nexeres API Spec
  version: 0.0.1
paths:
  /api/admin/orgs/:orgId/domains/:domain:
    post:
      consumes:
      - application/json
      description: Handles AdminUpdateDomain requests
      parameters:
      - description: Organization ID
        in: path
        name: orgId
        required: true
        type: string
      - description: Domain
        in: path
        name: domain
        required: true
        type: string
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/admin.AdminUpdateDomainRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.AdminUpdateDomainResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
      summary: AdminUpdateDomain Endpoint
      tags:
      - admin
...
  /api/admin/orgs/{orgId}/domains:
    put:
      consumes:
      - application/json
      description: Handles AdminCreateDomain requests
      parameters:
      - description: Organization ID
        in: path
        name: orgId
        required: true
        type: string
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/admin.AdminCreateDomainRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.AdminCreateDomainResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
      summary: AdminCreateDomain Endpoint
      tags:
      - admin
  /api/admin/orgs/{orgId}/domains/{domain}:
    delete:
      consumes:
      - application/json
      description: Handles AdminDeleteDomain requests
      parameters:
      - description: Organization ID
        in: path
        name: orgId
        required: true
        type: string
      - description: Domain
        in: path
        name: domain
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.AdminDeleteDomainResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
      summary: AdminDeleteDomain Endpoint
      tags:
      - admin
  /api/admin/orgs/{orgId}/domains/{domain}/verify:
    post:
      consumes:
      - application/json
      description: Handles AdminVerifyDomain requests
      parameters:
      - description: Organization ID
        in: path
        name: orgId
        required: true
        type: string
      - description: Domain to verify
        in: path
        name: domain
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.AdminVerifyDomainResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
      summary: AdminVerifyDomain Endpoint
      tags:
      - admin
  /api/admin/orgs/{orgId}/domains/{domain}/verify/code:
    get:
      consumes:
      - application/json
      description: Handles AdminGetDomainVerifyCode requests
      parameters:
      - description: Organization ID
        in: path
        name: orgId
        required: true
        type: string
      - description: Domain
        in: path
        name: domain
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.AdminGetDomainVerifyCodeResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
      summary: AdminGetDomainVerifyCode Endpoint
      tags:
      - admin
...
  /api/auth/verify-email/verify:
    post:
      consumes:
      - application/json
      description: Handles VerifyEmailToken requests
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.VerifyEmailTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.VerifyEmailTokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/reserr.ErrorResponse'
      summary: VerifyEmailToken Endpoint
      tags:
      - auth
...
securityDefinitions:
  APIKeyAuth:
    in: header
    name: X-NEXERES-API-Key
    type: apiKey
  AdminHeaderAuth:
    in: header
    name: X-NEXERES-Admin-Token
    type: apiKey
  RefreshHeaderAuth:
    in: header
    name: X-NEXERES-Refresh-Token
    type: apiKey
  SessionHeaderAuth:
    in: header
    name: X-NEXERES-Session-Token
    type: apiKey
swagger: "2.0"